test(settings-ui): share the mount harness across suites
Three suites carried their own copy of the createRoot/act mount
boilerplate. Move one harness into test/helpers and use it everywhere.

The native-fields suite also still built its props from the removed
SettingsFieldComponentProps shape, which only ran because jest does not
typecheck and the component ignores the extra props. It now uses the
internal NativeSettingsFieldProps contract.

Refs WOOPRD-3571
